Solve write ambiguity error on Apollo3 v2.1.0

+++ src/SparkFun_Qwiic_Humidity_AHT20.cpp
@@ -121,8 +121,8 @@ bool AHT20::initialize()
 {
     _i2cPort->beginTransmission(_deviceAddress);
     _i2cPort->write(sfe_aht20_reg_initialize);
-    _i2cPort->write(0x80);
-    _i2cPort->write(0x00);
+    _i2cPort->write((uint8_t)0x80);
+    _i2cPort->write((uint8_t)0x00);
     if (_i2cPort->endTransmission() == 0)
         return true;
     return false;
@@ -132,8 +132,8 @@ bool AHT20::triggerMeasurement()
 {
     _i2cPort->beginTransmission(_deviceAddress);
     _i2cPort->write(sfe_aht20_reg_measure);
-    _i2cPort->write(0x33);
-    _i2cPort->write(0x00);
+    _i2cPort->write((uint8_t)0x33);
+    _i2cPort->write((uint8_t)0x00);
     if (_i2cPort->endTransmission() == 0)
         return true;
     return false;
